refactor (step4) : 패키지 위치 변경 및 메소드 명칭 변경
